Direct3D 9 samples updated to use VS 2019, D3DX NuGet, etc.

+++ C++/Direct3D/ShadowVolume/ShadowVolume.cpp
@@ -1,7 +1,8 @@
 //--------------------------------------------------------------------------------------
 // File: ShadowVolume.cpp
 //
-// Copyright (c) Microsoft Corporation. All rights reserved.
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License (MIT).
 //--------------------------------------------------------------------------------------
 #include "DXUT.h"
 #include "DXUTcamera.h"
@@ -373,9 +374,9 @@ HRESULT GenerateShadowMesh( IDirect3DDevice9* pd3dDevice, ID3DXMesh* pMesh, ID3D
                         int nIndex;
                         int nVertIndex[3] =
                         {
-                            pdwPtRep[pdwIBData[f * 3]],
-                            pdwPtRep[pdwIBData[f * 3 + 1]],
-                            pdwPtRep[pdwIBData[f * 3 + 2]]
+                            static_cast<int>(pdwPtRep[pdwIBData[f * 3]]),
+                            static_cast<int>(pdwPtRep[pdwIBData[f * 3 + 1]]),
+                            static_cast<int>(pdwPtRep[pdwIBData[f * 3 + 2]])
                         };
                         nIndex = FindEdgeInMappingTable( nVertIndex[0], nVertIndex[1], pMapping, dwNumEdges );
